refactor(web): 추천 모임 섹션 초기 렌더에서 분리
